+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* ✅ test(slices-union): a new test case updated
 * 🐛 fix(slices-union): different types allowed
 * ✅ test(slices-union): 'union' function's test cases and benchmarks added
 * 🎉 feat(slices-union): 'union' function added